refactor(list): split renumber.lua into focused sub-modules (#274)
